Move queue with sizes to abstract-ops
