+++ src/lib/abstract-ops/queue-with-sizes.ts
@@ -1,6 +1,6 @@
-import assert from '../stub/assert';
-import { SimpleQueue } from './simple-queue';
-import { IsFiniteNonNegativeNumber } from './abstract-ops/miscellaneous';
+import assert from '../../stub/assert';
+import { SimpleQueue } from '../simple-queue';
+import { IsFiniteNonNegativeNumber } from './miscellaneous';
 
 export interface QueueContainer<T> {
   _queue: SimpleQueue<T>;
